% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators_indirect.R
\name{indirect_weightit}
\alias{indirect_weightit}
\title{WeightIt Estimator for Indirect Standardization}
\usage{
indirect_weightit(
  data,
  trt,
  outcome,
  baseline,
  outcome_type = "binomial",
  method = "glm",
  verbose = FALSE,
  control = standardization_control()
)
}
\arguments{
\item{data}{\[\code{data.frame}\]\cr
A \code{data.frame} containing all baseline, treatment, and outcome variables.}

\item{trt}{\[\code{character}\]\cr
Column name of treatment variable.}

\item{outcome}{\[\code{character}\]\cr
Column name of outcome variable.}

\item{baseline}{\[\code{character}\]\cr
Vector of column names of baseline variables.}

\item{outcome_type}{\[\code{character}\]\cr
Outcome variable type: binomial (binary) or continuous.}

\item{method}{\[\code{character}\]\cr
Method used within WeightIt for weights estimation. See WeightIt documentation for available methods.}

\item{verbose}{\[\code{logical}]\cr
Whether to print information messages during fitting}

\item{control}{\[\code{standardization_control}\]\cr
Additional tuning parameters for controlling fitting. Specify using \link{standardization_control}.}
}
\value{
A list of class \code{smr}
}
\description{
WeightIt Estimator for Indirect Standardization
}
