% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_providers}
\alias{simulate_providers}
\title{Simulate test data for provider profiling}
\usage{
simulate_providers(
  N = 1000,
  providers = 5,
  covariates = 5,
  seed = NA,
  effect_seed = NA
)
}
\arguments{
\item{N}{number of observations}

\item{providers}{number of providers}

\item{covariates}{number of baseline covariates}

\item{seed}{random seed for generating provider effects}

\item{effect_seed}{random seed for generating observations}
}
\description{
Simulate test data for provider profiling
}
