% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{standardization_control}
\alias{standardization_control}
\title{Algorithm tuning parameters for standardization parameters}
\usage{
standardization_control(
  .return_full_fits = FALSE,
  .learners_trt_folds = 5,
  .learners_outcome_folds = 5
)
}
\arguments{
\item{.return_full_fits}{\[\code{logical(1)}\]\cr
Whether to return full SuperLearner fits}

\item{.learners_trt_folds}{\[\code{integer(1)}\]\cr
Number of cross-fitting folds to use in SuperLearner for estimating treatment assignment mechanism}

\item{.learners_outcome_folds}{\[\code{integer(1)}\]\cr
Number of cross-fitting folds to use in SuperLearner for estimating outcome regression.}
}
\description{
Algorithm tuning parameters for standardization parameters
}
